+++ src/Layers/TranslationCreditsLayer.hpp
@@ -17,13 +17,17 @@ class TranslationCreditsLayer : public SillyBaseLayer
         CCSprite* background;
         ButtonSprite* sprUse;
         CCMenuItemSpriteExtra* btnUse;
+        int killCount = 0;
 
         virtual void customSetup();
 
         void onPlayerProfile(CCObject* sender);
         void onKill(CCObject* sender);
-
         void onUse(CCObject* sender);
+
+        void spawnBigParticle();
+        void spawnMediumParticle();
+        void spawnSmallParticle();
 
         static TranslationCreditsLayer* create(matjson::Value language, std::filesystem::path path);
         static TranslationCreditsLayer* addToScene(matjson::Value language, std::filesystem::path path);