+++ changelog.md
@@ -1,6 +1,7 @@
 # 1.7.7
 
 - Added **Localization / Translations**!!!
+-- The language can be changed in the **Config** tab of the mod menu
 
 # 1.7.6